Implement per-user VLESS fwmarks for freedom
Add uint32 fwmark support to static and dynamic VLESS users, propagate authenticated marks to direct Freedom sockets, and fail closed for unsafe mux and pooled transport paths.

+++ proxy/dokodemo/dokodemo.go
@@ -179,7 +179,7 @@ func (d *DokodemoDoor) Process(ctx context.Context, network net.Network, conn st
 			}
 			var mark int
 			if d.sockopt != nil {
-				mark = int(d.sockopt.Mark)
+				mark = int(int32(d.sockopt.Mark))
 			}
 			pConn, err := FakeUDP(addr, mark)
 			if err != nil {
